kills start menu

+++ modules/win_scheduled_tasks/files/at_task_user_logon.ps1
@@ -104,7 +104,13 @@ switch ($os_version) {
             }
         }
         ## Prevent start menu from popping up during tests
-        Get-Process -Name StartMenuExperienceHost | Stop-Process -Force
+        while ($true) {
+            start-sleep -s 5
+            $start_menu = (Get-Process -Name StartMenuExperienceHost -ErrorAction SilentlyContinue)
+            if ($start_menu) {
+                Get-Process -Name StartMenuExperienceHost | Stop-Process -Force
+            }
+        }
     }
     "win_2012" {
         ## Ensure strong encryption
